deprecate some auto_parallel uts
