+++ test/auto_parallel/CMakeLists.txt
@@ -214,26 +214,15 @@ if(WITH_DISTRIBUTE AND WITH_GPU)
   # End of unittests WITH multi cards and WITHOUT timeout
 
   # NOTE(zyl): unittests WITH single card and timeout
-  py_test_modules(test_parallel_tuner MODULES test_parallel_tuner)
-  set_tests_properties(test_parallel_tuner PROPERTIES TIMEOUT 120)
-  py_test_modules(test_parallel_tuner_full MODULES test_parallel_tuner_full)
-  set_tests_properties(test_parallel_tuner_full PROPERTIES TIMEOUT 120)
-  py_test_modules(test_parallel_tuner_predict MODULES
-                  test_parallel_tuner_predict)
-  set_tests_properties(test_parallel_tuner_predict PROPERTIES TIMEOUT 120)
   py_test_modules(test_tuning_recompute MODULES test_tuning_recompute)
   set_tests_properties(test_tuning_recompute PROPERTIES TIMEOUT 300)
   py_test_modules(test_tuning_recompute_with_amp MODULES
                   test_tuning_recompute_with_amp)
   set_tests_properties(test_tuning_recompute_with_amp PROPERTIES TIMEOUT 60)
-  py_test_modules(test_fused_linear_pass MODULES test_fused_linear_pass)
-  set_tests_properties(test_fused_linear_pass PROPERTIES TIMEOUT 40)
   py_test_modules(test_align_tool MODULES test_align_tool)
   set_tests_properties(test_align_tool PROPERTIES TIMEOUT 20)
   py_test_modules(test_pass_base_list MODULES test_pass_base_list)
   set_tests_properties(test_pass_base_list PROPERTIES TIMEOUT 40)
-  py_test_modules(test_fuse_adamw_pass MODULES test_fuse_adamw_pass)
-  set_tests_properties(test_fuse_adamw_pass PROPERTIES TIMEOUT 20)
   py_test_modules(test_rule_based_tuner_o2 MODULES test_rule_based_tuner_o2)
   set_tests_properties(test_rule_based_tuner_o2 PROPERTIES TIMEOUT 50)
   py_test_modules(test_semi_auto_parallel_functional_in_single_card MODULES
@@ -246,7 +235,6 @@ if(WITH_DISTRIBUTE AND WITH_GPU)
   py_test_modules(test_tunable_space MODULES test_tunable_space)
   py_test_modules(test_recorder MODULES test_recorder)
   py_test_modules(test_trial MODULES test_trial)
-  py_test_modules(test_new_cost_model MODULES test_new_cost_model)
   py_test_modules(test_dist_reshape MODULES test_dist_reshape)
   py_test_modules(test_dist_pnorm MODULES test_dist_pnorm)
   py_test_modules(test_dist_embedding MODULES test_dist_embedding)
@@ -262,19 +250,13 @@ if(WITH_DISTRIBUTE AND WITH_GPU)
   py_test_modules(test_cluster_v2 MODULES test_cluster_v2)
   py_test_modules(test_process_mesh_v2 MODULES test_process_mesh_v2)
   py_test_modules(test_dist_attr_v2 MODULES test_dist_attr_v2)
-  py_test_modules(test_serialization MODULES test_serialization)
-  py_test_modules(test_lr_grad_clip MODULES test_lr_grad_clip)
   py_test_modules(test_dist_matmul MODULES test_dist_matmul)
-  py_test_modules(test_process_mesh MODULES test_process_mesh)
-  py_test_modules(test_interface MODULES test_interface)
   py_test_modules(test_strategy MODULES test_strategy)
   py_test_modules(test_dist_shape MODULES test_dist_shape)
   py_test_modules(test_dist_assign MODULES test_dist_assign)
   py_test_modules(test_conditional_block_reshard MODULES
                   test_conditional_block_reshard)
   py_test_modules(test_engine_api_error MODULES test_engine_api_error)
-  py_test_modules(test_fp16_assign MODULES test_fp16_assign)
-  py_test_modules(test_group_operators MODULES test_group_operators)
   py_test_modules(test_cluster_partition MODULES test_cluster_partition)
   py_test_modules(test_convert_to_process_meshes MODULES
                   test_convert_to_process_meshes)
