+++ test/deprecated/auto_parallel/test_fp16_assign_deprecated.py
