+++ test/deprecated/auto_parallel/test_lr_grad_clip_deprecated.py
@@ -12,7 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import sys
 import unittest
+
+sys.path.append("../../auto_parallel")
 
 from test_to_static import MLPLayer, MyDataset
