+++ test/deprecated/auto_parallel/test_parallel_tuner_full_deprecated.py
@@ -13,8 +13,10 @@
 # limitations under the License.
 
 
+import sys
 import unittest
 
+sys.path.append("../../legacy_test")
 import auto_parallel_gpt_model as modeling
 from auto_parallel_gpt_model import (
     GPTForPretraining,
@@ -31,9 +33,11 @@ from paddle.distributed.auto_parallel.static.dist_context import (
     DistributedContext,
     set_default_distributed_context,
 )
+from paddle.distributed.auto_parallel.static.planner_v2 import Planner
 from paddle.distributed.auto_parallel.static.tuner.parallel_tuner import (
     ParallelTuner,
 )
+from paddle.distributed.auto_parallel.strategy import Strategy
 
 paddle.enable_static()
 
@@ -60,11 +64,11 @@ def get_program_v3():
     train_program = static.Program()
     start_program = static.Program()
     modeling.init_global()
-    modeling._global_parallel_strategy = None
-    # modeling.DPMPPP_MESH_LIST = [
-    #     ProcessMesh([[0, 1], [2, 3]], dim_names=["x", "y"]),
-    #     ProcessMesh([[4, 5], [6, 7]], dim_names=["x", "y"])
-    # ]
+    modeling._global_parallel_strategy = "dp_mp_pp"
+    modeling.DPMPPP_MESH_LIST = [
+        ProcessMesh([[0, 1], [2, 3]], dim_names=["x", "y"]),
+        ProcessMesh([[4, 5], [6, 7]], dim_names=["x", "y"]),
+    ]
     with static.program_guard(train_program, start_program):
         tokens = paddle.static.data(
             name="tokens", shape=[batch_size, sequence_len], dtype='int64'
@@ -101,7 +105,7 @@ def get_program_v3():
             eos_token_id=7,
             bos_token_id=0,
             eol_token_id=3,
-            pp_degree=1,
+            pp_degree=len(modeling.DPMPPP_MESH_LIST),
         )
 
         model = GPTForPretraining(
@@ -136,8 +140,8 @@ def get_program_v3():
     )
 
 
-class TestParallelTunerTrain(unittest.TestCase):
-    def test_tune_with_train(self):
+class TestParallelTunerFull(unittest.TestCase):
+    def test_tune_with_planner(self):
         flag = False
         set_default_distributed_context(DistributedContext())
         (
@@ -151,6 +155,8 @@ class TestParallelTunerTrain(unittest.TestCase):
         ) = get_program_v3()
         cluster = Cluster()
         cluster.gen_default_config_cluster(node_count=1, device_count=8)
+        strategy = Strategy()
+        strategy.auto_mode = "full_random"
         dist_context = DistributedContext(
             train_program,
             start_program,
@@ -159,11 +165,14 @@ class TestParallelTunerTrain(unittest.TestCase):
             feed_vars,
             fetch_vars,
             cluster,
+            strategy,
         )
         dist_context.initialize()
-        parallel_tuner = ParallelTuner(dist_context, max_trials=3, mode="train")
-        parallel_tuner.tune()
-        parallel_tuner._store_best_parallel_strategy()
+        planner = Planner("train", dist_context)
+        planner._parallel_tuner = ParallelTuner(
+            planner._dist_context, mode=planner._mode, max_trials=3
+        )
+        planner.plan()
         flag = True
         self.assertTrue(flag)
