+++ test/deprecated/auto_parallel/test_parallel_tuner_predict_deprecated.py
@@ -13,7 +13,10 @@
 # limitations under the License.
 
 
+import sys
 import unittest
+
+sys.path.append("../../legacy_test")
 
 import auto_parallel_gpt_model as modeling
 from auto_parallel_gpt_model import (
