+++ tools/gpups_test.sh
@@ -103,7 +103,7 @@ parallel_list="^init_phi_test$|\
 ^test_fused_gemm_epilogue_op_with_es$|\
 ^test_fused_layernorm_residual_dropout_bias$|\
 ^test_fused_linear_param_grad_add$|\
-^test_fused_linear_pass$|\
+^test_fused_linear_pass_deprecated$|\
 ^test_fused_matmul_bias$|\
 ^test_fused_multi_transformer_decoder_pass$|\
 ^test_fused_multi_transformer_encoder_pass$|\
